Suggestion : NumericFixedInterval should be able to take doubles and not just integers
This commit modifies TickGenerators.NumericFixedInterval so that the intervals can be doubles.

+++ src/ScottPlot5/ScottPlot5/TickGenerators/NumericFixedInterval.cs
@@ -1,6 +1,6 @@
 ﻿namespace ScottPlot.TickGenerators;
 
-public class NumericFixedInterval(int interval = 1) : ITickGenerator
+public class NumericFixedInterval(double interval = 1) : ITickGenerator
 {
     public double Interval { get; set; } = interval;
